moves read subscription to the internal package
